// src/generated/source-data.ts
export const gripeText = [
  {
    "text": "\r\nYou are not permitted to GRIPE\r\nwhile under RED alert!\r\n",
    "file": "WARMAC.MAC",
    "line": 4719
  },
  {
    "text": "Enter gripe, end with ^Z\r\n",
    "file": "WARMAC.MAC",
    "line": 4733
  },
  {
    "text": "[Only 2 more message lines allowed]\r\n",
    "file": "WARMAC.MAC",
    "line": 4759
  },
  {
    "text": "[Too many lines -- end of gripe]\r\n",
    "file": "WARMAC.MAC",
    "line": 4761
  },
  {
    "text": "**** Data out of bounds ****\r\n**** Command line:\r\n",
    "file": "WARMAC.MAC",
    "line": 4774
  },
  {
    "text": "*** PDL: ",
    "file": "WARMAC.MAC",
    "line": 4845
  },
  {
    "text": "*** HITQL-1:",
    "file": "WARMAC.MAC",
    "line": 4865
  },
  {
    "text": "*** LOKTAB:",
    "file": "WARMAC.MAC",
    "line": 4881
  },
  {
    "text": "----------",
    "file": "WARMAC.MAC",
    "line": 4920
  },
  {
    "text": "%Can't write DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4932
  },
  {
    "text": "%DECWAR.GRP being modified; trying again\r\n",
    "file": "WARMAC.MAC",
    "line": 4933
  },
  {
    "text": "%Can't get core to read DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4951
  },
  {
    "text": "%Can't read DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4958
  },
  {
    "text": "%Can't write DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4969
  },
  {
    "text": "%Can't get more core\r\n",
    "file": "WARMAC.MAC",
    "line": 4992
  }
] as const;